# docs/openapi3.yaml
openapi: 3.0.0
...
info:
  contact:
    email: shpakovskiipa@gmail.com
    name: API Support
  description: This is a Eschool API.
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: Eschool API
  version: "1.0"
paths:
  /auth/logout:
    post:
      description: logout
      responses:
        "200":
          description: message
          content:
            application/json:
              schema:
                type: string
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/internal_adapter_delivery_http_v1.RestErrorUnauthor\
                  ized"
        "500":
          description: Internal Server Error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/internal_adapter_delivery_http_v1.RestErrorInternal\
                  Error"
      summary: Logout
      tags:
        - auth
  /auth/refresh:
    post:
      description: refresh
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/github_com_paw1a_eschool_internal_adapter_delivery_\
                http_v1_dto.RefreshDTO"
        description: refresh payload
        required: true
      responses:
        "200":
          description: auth token
          content:
            application/json:
              schema:
                type: string
        "400":
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/internal_adapter_delivery_http_v1.RestErrorBadReque\
                  st"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/internal_adapter_delivery_http_v1.RestErrorUnauthor\
                  ized"
        "500":
          description: Internal Server Error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/internal_adapter_delivery_http_v1.RestErrorInternal\
                  Error"
      summary: TokenRefresh
      tags:
        - auth
  /auth/sign-in:
    post:
      description: sign-in
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/github_com_paw1a_eschool_internal_adapter_delivery_\
                http_v1_dto.SignInDTO"
        description: credentials
        required: true
      responses:
        "200":
          description: auth token
          content:
            application/json:
              schema:
                type: string
        "400":
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/internal_adapter_delivery_http_v1.RestErrorBadReque\
                  st"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/internal_adapter_delivery_http_v1.RestErrorUnauthor\
                  ized"
        "500":
          description: Internal Server Error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/internal_adapter_delivery_http_v1.RestErrorInternal\
                  Error"
      summary: SignIn
      tags:
        - auth
  /auth/sign-up:
    post:
      description: sign-up
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/github_com_paw1a_eschool_internal_adapter_delivery_\
                http_v1_dto.SignUpDTO"
        description: user information
        required: true
      responses:
        "201":
          description: message
          content:
            application/json:
              schema:
                type: string
        "400":
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/internal_adapter_delivery_http_v1.RestErrorBadReque\
                  st"
        "409":
          description: Conflict
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/internal_adapter_delivery_http_v1.RestErrorConflict"
        "500":
          description: Internal Server Error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/internal_adapter_delivery_http_v1.RestErrorInternal\
                  Error"
      summary: SignUp
      tags:
        - auth
...
servers:
  - url: //localhost:8080/api/v1
components:
  securitySchemes:
    ApiKeyAuth:
      in: header
      name: Authorization
      type: apiKey
  schemas:
...
    internal_adapter_delivery_http_v1.RestErrorConflict:
      properties:
        error:
          example: conflict
          type: string
        status:
          example: 409
          type: integer
        timestamp:
          example: 2020-11-10T23:00:00+00:00
          type: string
      type: object
...